Trait selector kernel test revisions
